hotfix nr 2

+++ digital_twin/python/subscriber.py
@@ -19,9 +19,9 @@ def on_message(client, userdata, msg):
     logging.info(msg.topic + ": "+ msg.payload.decode())
     print(msg.topic + ": "+ msg.payload.decode())
     
-    if (int(msg.payload.decode()) < LOWERBOUND):
+    if (float(msg.payload.decode()) < LOWERBOUND):
         client.publish(MQTT_PATH_PUBLISH, payload = "HEATER: ON", qos = 0, retain=False)
-    elif (int(msg.payload.decode()) > HIGHERBOUND):
+    elif (float(msg.payload.decode()) > HIGHERBOUND):
         client.publish(MQTT_PATH_PUBLISH, payload = "HEATER: OFF", qos = 0, retain=False)
